refactor(CLOUDDEV-625): Refactor flavor utils
